packages/sdk-py/pyproject.toml
[project]
name = "signet-verify"
version = "0.1.0"
description = "Signet verification SDK (Python)"
requires-python = ">=3.12"
dependencies = ["rfc8785>=0.1.0", "PyNaCl>=1.5.0"]

[project.optional-dependencies]
dev = ["pytest>=8.3.3", "pytest-cov>=5.0.0"]

[build-system]
requires = ["setuptools>=69.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["signet_verify"]
